+++ vlm/src/test/scala/geotrellis/contrib/vlm/spark/RasterSummarySpec.scala
@@ -52,7 +52,7 @@ class RasterSummarySpec extends FunSpec with TestEnvironment with BetterRasterMa
     }
 
     it("should collect summary for a tiled to layout source") {
-      val inputPath = Resource.path("img/aspect-tiled.tif")
+      val inputPath = GeoTiffDataPath(Resource.path("img/aspect-tiled.tif"))
       val files = inputPath :: Nil
       val targetCRS = WebMercator
       val method = Bilinear
@@ -92,7 +92,7 @@ class RasterSummarySpec extends FunSpec with TestEnvironment with BetterRasterMa
   }
 
   it("should create ContextRDD from RDD of GeoTiffRasterSources") {
-    val inputPath = Resource.path("img/aspect-tiled.tif")
+    val inputPath = GeoTiffDataPath(Resource.path("img/aspect-tiled.tif"))
     val files = inputPath :: Nil
     val targetCRS = WebMercator
     val method = Bilinear